{{ cookiecutter.organization }}-{{ cookiecutter.project_name }}/backend/pyproject.toml
[project]
name = "{{ cookiecutter.__python_package }}"
version = "1.0.0.dev0"
description = "{{ cookiecutter.title }}"
authors = [{name = "{{ cookiecutter.author }}", email = "{{ cookiecutter.author_email }}"}]
requires-python = ">={{ cookiecutter.python_version }}"
license = {text = "GPL-2.0-only"}
dependencies = [
    "plone.distribution",
    "plone.volto",
]

[project.optional-dependencies]
test = [
    "plone.app.contenttypes[test]",
    "pytest-plone",
]
production = [
    "relstorage[postgresql]",
    "psycopg[binary]",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = [
    "A",    # flake8-builtins (shadowing builtins)
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "PTH",  # flake8-use-pathlib
    "RUF",  # ruff-specific
    "SIM",  # flake8-simplify
    "T20",  # flake8-print
    "UP",   # pyupgrade
    "W",    # pycodestyle warnings
]

[tool.ruff.lint.isort]
force-single-line = true
from-first = true
lines-after-imports = 2
lines-between-types = 1
no-sections = true
order-by-type = false

[tool.ruff.lint.pycodestyle]
max-line-length = 120
